R simple program to find the area : For square and rectangle.
Writing by Mékis AZONDEKON

DESCRIPTION
This is just a simple program that takes two arguments. It is very funny and awesome to find the areas of square and rectangle when you use. Work being automatically easy and quick.
The function’s name is Surface.
Surface takes two arguments, one is a and another b

The things you have to is : 
-Just replace the arguments that i have already seted as a and b ;
**In the case of square a=b, so you will supply the values 
**In the case of rectangle a the small side and b the other one, and the same work must be done too.