R skills
In this project the main point is to build a function that organize numbers (common numerics in R). As we complete the learning of R language, we think about the way of this project and how it will be done.
The following functions have done to solve that problems. It makes quickly task as organize numerics to the smallest to greatest and to greatest to smallest. The functions are : crois and decrois.
What is crois ?
‘crois’ is the function i’ve built in R to organize numbers to the smallest one to greatest. ‘crois’ takes four (4) arguments : a, b, c and d.
How to use ‘crois’ :
*Open your R studio
*Run ‘crois’ in your R studio console : You can highlight the script and copy it in your R studio console (make sure you hit Enter after that), or you can highlight it and click Run.
*Then you can use the function (i wonder you know how to do it).
What is ‘decrois’ ?
‘decrois’ is the function i’ve built in R to to organize numbers to greatest to smallest. ‘decrois’ takes four (4) arguments : a, b, c and d.
How to use ‘decrois’ :
How to use ‘crois’ :
*Open your R studio
*Run ‘decrois’ in your R studio console : You can highlight the script and copy it in your R studio console (make sure you hit Enter after that), or you can highlight it and click Run. *Then you can use the function (i wonder you know how to do it).